test: move the tests to correct module
